Refactor function of localization
